+++ src/lib.rs
@@ -22,9 +22,15 @@ This trait adds the arbitrary, match-based trimming methods to `&[u8]`, `Vec<u8>
 
 | Method | Description |
 | ------ | ----------- |
-| `trim_matches` | Trim arbitrary leading and trailing bytes via callback. |
-| `trim_start_matches` | Trim arbitrary leading bytes via callback. |
-| `trim_end_matches` | Trim arbitrary trailing bytes via callback. |
+| `trim_matches` | Trim arbitrary leading and trailing bytes. |
+| `trim_start_matches` | Trim arbitrary leading bytes. |
+| `trim_end_matches` | Trim arbitrary trailing bytes. |
+
+Each of these match methods accept either:
+* A single `u8`;
+* An array or slice of `u8`;
+* A `&BtreeSet<u8>`
+* A custom callback with signature `Fn(u8) -> bool`
 
 
 ### [`TrimMut`]
@@ -44,9 +50,17 @@ This trait brings _mutable_ match-based trimming `String`, `Vec<u8>`, and `Box<[
 
 | Method | Description |
 | ------ | ----------- |
-| `trim_matches_mut` | Trim arbitrary leading and trailing bytes via callback (mutably). |
-| `trim_start_matches_mut` | Trim arbitrary leading bytes via callback (mutably). |
-| `trim_end_matches_mut` | Trim arbitrary trailing bytes via callback (mutably). |
+| `trim_matches_mut` | Trim arbitrary leading and trailing bytes (mutably). |
+| `trim_start_matches_mut` | Trim arbitrary leading bytes (mutably). |
+| `trim_end_matches_mut` | Trim arbitrary trailing bytes (mutably). |
+
+Each of these match methods accept either:
+* A single T;
+* An array or slice of T;
+* A `&BtreeSet<T>`
+* A custom callback with signature `Fn(T) -> bool`
+
+Where T is `char` for `String`, and `u8` for `Vec<u8>`/`Box<[u8]>`.
 
 
 
@@ -65,7 +79,7 @@ This trait is implemented for `&[u8]`, `&str`, and `Iterator`s with `u8`/`char` 
 #![forbid(unsafe_code)]
 
 #![deny(
-	// TODO: clippy::allow_attributes_without_reason,
+	clippy::allow_attributes_without_reason,
 	clippy::correctness,
 	unreachable_pub,
 )]
@@ -77,7 +91,7 @@ This trait is implemented for `&[u8]`, `&str`, and `Iterator`s with `u8`/`char` 
 	clippy::perf,
 	clippy::style,
 
-	// TODO: clippy::allow_attributes,
+	clippy::allow_attributes,
 	clippy::clone_on_ref_ptr,
 	clippy::create_dir,
 	clippy::filetype_is_file,
@@ -111,13 +125,14 @@ This trait is implemented for `&[u8]`, `&str`, and `Iterator`s with `u8`/`char` 
 	unused_import_braces,
 )]
 
-#![allow(clippy::module_name_repetitions)] // Repetition is preferred.
+#![expect(clippy::module_name_repetitions, reason = "Repetition is preferred.")]
 
 #![no_std]
 
 extern crate alloc;
 
 mod iter;
+mod pattern;
 mod trim_mut;
 mod trim_slice;
 
@@ -130,7 +145,7 @@ pub use trim_slice::TrimSliceMatches;
 
 
 
-#[allow(clippy::trivially_copy_pass_by_ref)] // This signature is required.
+#[expect(clippy::trivially_copy_pass_by_ref, reason = "This signature is required.")]
 #[inline]
 /// # Not Whitespace.
 ///